Include copyright notice; separate out original content.
